+++ scripts/macros/session-turnover.js
@@ -33,32 +33,35 @@ async function endSession () {
       ? Number(thisSession) + 1
       : thisSession
 
-    const dialog = new Dialog({
-      title: game.i18n.localize("GMTOOLKIT.Dialog.SessionTurnover.UpdateSessionID.Title"),
+    foundry.applications.api.DialogV2.wait({
+      window: { title: game.i18n.localize("GMTOOLKIT.Dialog.SessionTurnover.UpdateSessionID.Title") },
+      rejectClose: false,
       content: `<form>
             <p>${game.i18n.format("GMTOOLKIT.Dialog.SessionTurnover.UpdateSessionID.CurrentSession", { thisSession })}</p>
             <div class="form-group">
             <p>${game.i18n.localize("GMTOOLKIT.Dialog.SessionTurnover.UpdateSessionID.NextSession")}</p> 
-            <input type="text" id="next-session" name="next-session" value="${nextSession}" />
+            <input type="text" id="nextsession" name="nextsession" value="${nextSession}" />
             </div>
             </form>`,
-      buttons: {
-        yes: {
+      buttons: [
+        {
           icon: "<i class='fas fa-check'></i>",
           label: game.i18n.localize("GMTOOLKIT.Dialog.Apply"),
-          callback: html => {
-            nextSession = html.find("#next-session").val()
+          action: "apply",
+          callback: (event, button, dialog) => {
+            result = new FormDataExtended(button.form).object
+            nextSession = result.nextsession
             game.settings.set("wfrp4e-gm-toolkit", "sessionID", nextSession)
             game.gmtoolkit.module.log(true, `Previous Session ID was ${thisSession}. Next Session ID is ${nextSession}.`)
           }
         },
-        no: {
+        {
           icon: "<i class='fas fa-times'></i>",
-          label: game.i18n.localize("GMTOOLKIT.Dialog.Cancel")
+          label: game.i18n.localize("GMTOOLKIT.Dialog.Cancel"),
+          action: "cancel"
         }
-      },
-      default: "yes"
-    }).render(true)
+      ]
+    })
   } // End Session ID Update
 
   game.gmtoolkit.module.log(false, "Completed Session Turnover tasks.")
@@ -66,8 +69,8 @@ async function endSession () {
 
 /* ==========
  * MACRO: Session Turnover
- * VERSION: 6.0.1
- * UPDATED: 2022-09-16
+ * VERSION: 8.1.0
+ * UPDATED: 2025-03-24
  * DESCRIPTION: Unified macro to run start and end of session admin, including awarding Experience Points, resetting Fortune, pausing the game and exporting the chat log.
  * TIP: Various default options can be defined in Session Management Settings under Module Settings.
  ========== */