+++ CHANGELOG.md
@@ -18,6 +18,7 @@ See [Issue Backlog](../../issues) and [Roadmap](../../milestones).
 - *Changed* Add XP UI to use DialogV2. [#304](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/304)
   - Revised calculation of henchman XP award to use `Math.floor()`. XP continues to be rounded down.
   - Restrict XP amount to only accept numbers. Remove redundant translation strings.
+- *Changed* Session Turnover UI to use DialogV2. [#302](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/issues/302)
 
 ## [Version 8.0.1](https://github.com/Jagusti/fvtt-wfrp4e-gmtoolkit/releases/tag/v8.0.1)  (2024-09-21)
 - *Added* WFRP4e v8 compatibility for Check Conditions macro.